refactor: async collection style
