+++ View Utils/Async/AsyncContentView.swift
@@ -99,83 +99,3 @@ enum AsyncContentPhase<Value> {
     case completed(value: Value)
     case failed(error: Error)
 }
-
-
-/// The style used in ``AsyncContentView``.
-@MainActor
-public struct AsyncContentStyle: Sendable {
-    public typealias Retry = @Sendable () -> Void
-    
-    /// Create a style.
-    /// - Parameters:
-    ///   - loadingView: The loading view.
-    ///   - errorView: Render the error view using an error and a retry action.
-    public init(@ViewBuilder loadingView: @MainActor @escaping @Sendable () -> some View,
-                @ViewBuilder errorView: @MainActor @escaping @Sendable (Error, @escaping Retry) -> some View) {
-        self.loadingView = { () -> AnyView in AnyView(loadingView()) }
-        self.errorView = { error, retry in AnyView(errorView(error, retry)) }
-    }
-    
-    let loadingView: @MainActor @Sendable () -> AnyView
-    let errorView: @MainActor @Sendable (Error, @escaping Retry) -> AnyView
-}
-
-extension AsyncContentStyle {
-    public static let paged = AsyncContentStyle {
-        ProgressView {
-            Text("Loading", bundle: .module)
-        }
-    } errorView: { error, retry in
-        VStack {
-            Text("Loading Failed", bundle: .module)
-                .font(.title)
-                .fontWeight(.bold)
-            
-            Text(error.localizedDescription)
-                .font(.callout)
-                .padding(.bottom)
-                .multilineTextAlignment(.center)
-                .foregroundStyle(.secondary)
-            
-            Button {
-                retry()
-            } label: {
-                Text("Retry", bundle: .module)
-            }
-            .foregroundStyle(Color.accentColor)
-        }
-        .padding()
-    }
-    
-    public static let widget = AsyncContentStyle {
-        HStack {
-            Spacer()
-            ProgressView()
-            Spacer()
-        }
-        .listRowBackground(Color.clear)
-    } errorView: { error, retry in
-        HStack {
-            Spacer()
-            VStack {
-                if let errorDescription = (error as? LocalizedError)?.errorDescription {
-                    Text(errorDescription)
-                        .foregroundStyle(.secondary)
-                } else {
-                    Text("Loading Failed", bundle: .module)
-                        .foregroundStyle(.secondary)
-                }
-                
-                Button {
-                    retry()
-                } label: {
-                    Text("Retry", bundle: .module)
-                }
-                .foregroundStyle(Color.accentColor)
-            }
-            .font(.caption)
-            .multilineTextAlignment(.center)
-            Spacer()
-        }
-    }
-}
